admin products view

+++ www/database/migrations/2018_11_22_082655_create_products_table.php
@@ -19,7 +19,7 @@ class CreateProductsTable extends Migration
             $table->string('name', 50);
             $table->text('description');
             $table->string('short_disc', 150);
-            $table->float('price', 5, 2);
+            $table->integer('price');
             $table->boolean('is_active')->default('1');
             $table->timestamps();
         });
